1. Исправил название пакета 2. Локаничнее назвал метод reverse 3. Убрал ненужный else 4. Исправил условие в классе Example. Теперь если масив пустой, то return сработает
