+++ src/main/java/school/faang/Grouping_users_by_age/Main.java
@@ -1,0 +1,9 @@
+package school.faang.Grouping_users_by_age;
+
+public class Main {
+    public static void main(String[] args) {
+        int[] integers = {};
+
+        Example.reverse(integers);
+    }
+}
