--- src/main/java/school/faang/module1/Main.java
@@ -1,9 +1,0 @@
-package school.faang.module1;
-
-public class Main {
-    public static void main(String[] args) {
-        int[] integers = {1, 2, 3, 4, 5};
-
-        Example.reverseIt(integers);
-    }
-}
